[data][cornell] So many I don't know and yes, no
- which quite explain why we our bot is so e ... smart? ...
